[ref] tracking|main|test: code cleaning

+++ Test/trackingTest.cpp
@@ -19,9 +19,9 @@ class TrackingTest : public ::testing::Test {
 //Curvature method test
 TEST_F(TrackingTest, CurvatureCenter) {
   Tracking tracking("test");
-  EXPECT_EQ(tracking.curvatureCenter( Point3f(1, 1, 0.25*M_PI + 0.5*M_PI), Point3f(2, 2, 0.5*M_PI + 0.5*M_PI)), Point2f(0, 0));
-  EXPECT_EQ(tracking.curvatureCenter( Point3f(1, 1, 0.25*5*M_PI + 0.5*M_PI), Point3f(2, 2, 0.5*M_PI + 0.5*M_PI)), Point2f(0, 0));
-  EXPECT_EQ(tracking.curvatureCenter( Point3f(1, 1, -3*0.25*M_PI + 0.5*M_PI), Point3f(2, 2, 0.5*M_PI + 0.5*M_PI)), Point2f(0, 0));
+  EXPECT_EQ(tracking.curvatureCenter( Point3d(1, 1, 0.25*M_PI + 0.5*M_PI), Point3d(2, 2, 0.5*M_PI + 0.5*M_PI)), Point2d(0, 0));
+  EXPECT_EQ(tracking.curvatureCenter( Point3d(1, 1, 0.25*5*M_PI + 0.5*M_PI), Point3d(2, 2, 0.5*M_PI + 0.5*M_PI)), Point2d(0, 0));
+  EXPECT_EQ(tracking.curvatureCenter( Point3d(1, 1, -3*0.25*M_PI + 0.5*M_PI), Point3d(2, 2, 0.5*M_PI + 0.5*M_PI)), Point2d(0, 0));
 }
 
 
@@ -96,21 +96,21 @@ TEST_F(TrackingTest, AngleDifferrence) {
 TEST_F(TrackingTest, Reassignement) {
   Tracking tracking("test");
   
-  vector<Point3f> past, input;
+  vector<Point3d> past, input;
   vector<int> order;
   
-  past = { Point3f(1), Point3f(2), Point3f(3) };
-  input = { Point3f(3), Point3f(2), Point3f(1) };
+  past = { Point3d(1), Point3d(2), Point3d(3) };
+  input = { Point3d(3), Point3d(2), Point3d(1) };
   order = { 2, 1, 0 };
   EXPECT_EQ(tracking.reassignment(past, input, order), past);
 
-  past = { Point3f(1), Point3f(2), Point3f(3) };
-  input = { Point3f(3), Point3f(2) };
+  past = { Point3d(1), Point3d(2), Point3d(3) };
+  input = { Point3d(3), Point3d(2) };
   order = { -1, 1, 0 };
   EXPECT_EQ(tracking.reassignment(past, input, order), past);
 
-  past = { Point3f(1), Point3f(2), Point3f(3) };
-  input = { Point3f(3), Point3f(2), Point3f(4), Point3f(1) };
+  past = { Point3d(1), Point3d(2), Point3d(3) };
+  input = { Point3d(3), Point3d(2), Point3d(4), Point3d(1) };
   order = { 3, 1, 0 };
   EXPECT_EQ(tracking.reassignment(past, input, order), past);
 }
@@ -143,7 +143,7 @@ TEST_F(TrackingTest, InformationOrientation) {
   UMat imageReference, rotate;
   
   imread("../dataSet/assymetric_left.png", IMREAD_GRAYSCALE).copyTo(imageReference);
-  Point2f center(double(imageReference.cols)*.5, double(imageReference.rows)*.5);
+  Point2d center(double(imageReference.cols)*.5, double(imageReference.rows)*.5);
   test = tracking.objectInformation(imageReference);
   tracking.objectDirection(imageReference, Point(test.at(0), test.at(1)), test);
   EXPECT_EQ(round(test.at(2)*1000), round(1000*(M_PI)));
@@ -153,63 +153,63 @@ TEST_F(TrackingTest, InformationOrientation) {
 TEST_F(TrackingTest, costFunction) {
   Tracking tracking("test");
   
-  vector<Point3f> past, current;
+  vector<Point3d> past, current;
   vector<int> order, test;
 
   // Only translation
-  past = { Point3f(10, 10, 0), Point3f(30, 40, 0), Point3f(50, 60, 0) };
-  current = { Point3f(50, 60, 0), Point3f(10, 10, 0), Point3f(30, 40, 0) };
+  past = { Point3d(10, 10, 0), Point3d(30, 40, 0), Point3d(50, 60, 0) };
+  current = { Point3d(50, 60, 0), Point3d(10, 10, 0), Point3d(30, 40, 0) };
   order = tracking.costFunc(past, current, 10, 90, 1, 20);
   test = {1, 2, 0};
   EXPECT_EQ(order, test);
   EXPECT_EQ(tracking.reassignment(past, current, order), past);
 
   // Only translation
-  past = { Point3f(10, 10, 0), Point3f(30, 40, 0), Point3f(50, 60, 0) };
-  current = { Point3f(50, 50, 0), Point3f(10, 20, 0), Point3f(30, 30, 0) };
+  past = { Point3d(10, 10, 0), Point3d(30, 40, 0), Point3d(50, 60, 0) };
+  current = { Point3d(50, 50, 0), Point3d(10, 20, 0), Point3d(30, 30, 0) };
   order = tracking.costFunc(past, current, 10, 90, 1, 1);
   test = {1, 2, 0};
   EXPECT_NE(order, test);
 
   // Only angle no movement
-  past = { Point3f(10, 10, 90), Point3f(30, 40, 180), Point3f(50, 60, 0) };
-  current = { Point3f(50, 60, 0), Point3f(10, 10, 90), Point3f(30, 40, 180) };
+  past = { Point3d(10, 10, 90), Point3d(30, 40, 180), Point3d(50, 60, 0) };
+  current = { Point3d(50, 60, 0), Point3d(10, 10, 90), Point3d(30, 40, 180) };
   order = tracking.costFunc(past, current, 10, 90, 0, 20);
   test = {1, 2, 0};
   EXPECT_EQ(order, test);
   EXPECT_EQ(tracking.reassignment(past, current, order), past);
   
   // Only translation
-  past = { Point3f(10, 10, 0), Point3f(30, 40, 0), Point3f(50, 60, 0) };
-  current = { Point3f(60, 60, 0), Point3f(10, 0, 0), Point3f(30, 50, 0) };
+  past = { Point3d(10, 10, 0), Point3d(30, 40, 0), Point3d(50, 60, 0) };
+  current = { Point3d(60, 60, 0), Point3d(10, 0, 0), Point3d(30, 50, 0) };
   order = tracking.costFunc(past, current, 15, 90, 0.5, 20);
   test = {1, 2, 0};
   EXPECT_EQ(order, test);
 
   // Rotation and translation
-  past = { Point3f(10, 10, 0), Point3f(30, 40, 3*0.5*M_PI), Point3f(50, 60, 0.5*M_PI) };
-  current = { Point3f(60, 60, 0), Point3f(10, 0, 0.5*M_PI), Point3f(30, 50, 0) };
+  past = { Point3d(10, 10, 0), Point3d(30, 40, 3*0.5*M_PI), Point3d(50, 60, 0.5*M_PI) };
+  current = { Point3d(60, 60, 0), Point3d(10, 0, 0.5*M_PI), Point3d(30, 50, 0) };
   order = tracking.costFunc(past, current, 15, 90, 0.5, 20);
   test = {1, 2, 0};
   EXPECT_EQ(order, test);
 
   // Only translation
-  past = { Point3f(10, 10, 0), Point3f(30, 40, 0), Point3f(50, 60, 0) };
-  current = { Point3f(60, 60, 0), Point3f(10, 0, 0), Point3f(30, 50, 0), Point3f(30, 150, 0)};
+  past = { Point3d(10, 10, 0), Point3d(30, 40, 0), Point3d(50, 60, 0) };
+  current = { Point3d(60, 60, 0), Point3d(10, 0, 0), Point3d(30, 50, 0), Point3d(30, 150, 0)};
   order = tracking.costFunc(past, current, 15, 90, 0.5, 20);
   test = {1, 2, 0};
   EXPECT_EQ(order, test);
 
   // Rotation and translation
-  past = { Point3f(10, 10, 0), Point3f(30, 40, 3*0.5*M_PI), Point3f(50, 60, 0.5*M_PI) };
-  current = { Point3f(60, 60, 0), Point3f(10, 0, 0.5*M_PI) };
+  past = { Point3d(10, 10, 0), Point3d(30, 40, 3*0.5*M_PI), Point3d(50, 60, 0.5*M_PI) };
+  current = { Point3d(60, 60, 0), Point3d(10, 0, 0.5*M_PI) };
   order = tracking.costFunc(past, current, 15, 90, 0.5, 20);
   test = {1, -1, 0};
   EXPECT_EQ(order, test);
 
   // Only rotation
-  past = { Point3f(10, 10, 0), Point3f(30, 40, 3*0.5*M_PI), Point3f(50, 60, 0.5*M_PI) };
-  current = { Point3f(50, 60, 0), Point3f(10, 10, 0.5*M_PI), Point3f(30, 40, 0) };
+  past = { Point3d(10, 10, 0), Point3d(30, 40, 3*0.5*M_PI), Point3d(50, 60, 0.5*M_PI) };
+  current = { Point3d(50, 60, 0), Point3d(10, 10, 0.5*M_PI), Point3d(30, 40, 0) };
   order = tracking.costFunc(past, current, 15, 90, 0.5, 20);
   test = {1, 2, 0};
   EXPECT_EQ(order, test);
